Fixing paths in internallauncher for bundled exes
Making the internallauncher script check to see if the exe app bundle directory exists before setting the variable to that directory. This fixes the errors that were shown when trying to run visit from the distribution while maintaining the existing functionality from the installation directory.

This change was made so that correct pythonpath is set while creating a movie through composer in windows. In Windows, VSim distro has python in Contents directory. So if the path set by this code initially does not exist, try setting the path to the Contents directory

Add another location for the resources.