Fix issues with dynamic library ids on MacOS.
Make sure that visitmodule.so gets built with fully-qualified paths to dependent libraries. We expect this to be the case in the *Sim installers.

Change the install_name_dir of visitmodule to match the installed location instead of the build location

Set the install_name_dir of libraries to ensure that their id and linked library paths are set properly at install time